refactor(auth): move fwa leader role setter from /set to /fwa leader-role
